+++ src/app/ui/layout/logout-modal/logout-modal.test.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+import LogoutModal from './logout-modal.component';
+
+describe('Logout Modal', () => {
+    const mockOnLogout = jest.fn();
+    const mockOnCancel = jest.fn();
+    const isOnMobile = false;
+
+    test('calls onLogout and onCancel when buttons are clicked', () => {
+        render(
+            <LogoutModal
+                onLogout={mockOnLogout}
+                onCancel={mockOnCancel}
+                isOnMobile={isOnMobile}
+            />
+        );
+
+        const logoutButton = screen.getByText('Desconectar');
+        const cancelButton = screen.getByText('Cancelar');
+
+        // Simulate button clicks
+        logoutButton.click();
+        cancelButton.click();
+
+        expect(mockOnLogout).toHaveBeenCalledTimes(1);
+        expect(mockOnCancel).toHaveBeenCalledTimes(1);
+    });
+});
