+++ src/app/ui/layout/logout-modal/logout-modal.types.tsx
@@ -1,0 +1,5 @@
+export interface LogoutModalProps {
+    onLogout: () => void,
+    onCancel: () => void,
+    isOnMobile: boolean,
+};
