--- src/app/ui/layout/modal/modal.types.tsx
@@ -1,9 +1,0 @@
-export interface ModalProps {
-    onLogout: () => void,
-    onCancel: () => void,
-    modalTitle: string,
-    modalMessage: string,
-    logoutButtonType: "alert" | "primary" | "error",
-    cancelButtonType: "alert" | "primary" | "error",
-    isOnMobile: boolean,
-};
